+++ packages/ibiza-themes/src/azure/AzureThemeDark.ts
@@ -1,0 +1,17 @@
+import { createTheme, ITheme } from 'office-ui-fabric-react';
+import { CommonSemanticColors, DarkSemanticColors } from './AzureColors';
+
+export const AzureThemeDark: ITheme = createTheme({
+  semanticColors: {
+    bodyBackground: DarkSemanticColors.background,
+    bodyText: DarkSemanticColors.text.body,
+    disabledBackground: CommonSemanticColors.backgrounds.disabled,
+    disabledBodyText: DarkSemanticColors.text.disabledText,
+    errorBackground: DarkSemanticColors.controlOutlines.error,
+    errorText: DarkSemanticColors.text.error,
+    focusBorder: DarkSemanticColors.controlOutlines.accent,
+    inputBackground: DarkSemanticColors.background,
+    inputText: DarkSemanticColors.text.value,
+    inputBorder: CommonSemanticColors.textControlOutline.rest
+  }
+});
