+++ packages/ibiza-themes/src/azure/Constants.ts
@@ -1,5 +1,5 @@
 import { ScreenWidthMaxMedium, ScreenWidthMaxSmall, ScreenWidthMinMedium, getScreenSelector } from 'office-ui-fabric-react/lib/Styling';
 
-export const borderRadius = '2px';
+export const borderWidthError = '2px';
 export const MinimumScreenSelector = getScreenSelector(0, ScreenWidthMaxSmall);
 export const MediumScreenSelector = getScreenSelector(ScreenWidthMinMedium, ScreenWidthMaxMedium);
