--- packages/ibiza-themes/src/ibiza/IExtendedTheme.ts
@@ -1,7 +1,0 @@
-import { ITheme } from 'office-ui-fabric-react';
-import { IExtendedThemeColors } from './IExtendedThemeColors';
-
-export interface IExtendedTheme {
-  extendedColors: IExtendedThemeColors;
-  theme: ITheme;
-}
