--- packages/ibiza-themes/src/ibiza/index.ts
@@ -1,7 +1,0 @@
-export * from './IbizaColors';
-export * from './IbizaDepths';
-export * from './IbizaMotion';
-export * from './IbizaStyles';
-export * from './IbizaExtendedThemeDark';
-export * from './IbizaExtendedThemeLight';
-export * from './IbizaType';
